requirement.txt
Babel==1.3
Django==1.7.3
Pillow==2.7.0
argparse==1.2.1
django-storage-swift==1.2.3
futures==2.2.0
iso8601==0.1.10
netaddr==0.7.13
netifaces==0.10.4
oslo.config==1.6.0
oslo.i18n==1.3.0
oslo.serialization==1.2.0
oslo.utils==1.2.1
pbr==0.10.7
prettytable==0.7.2
python-keystoneclient==1.0.0
python-swiftclient==2.3.1
pytz==2014.10
requests==2.5.1
simplejson==3.6.5
six==1.9.0
stevedore==1.2.0
wsgiref==0.1.2
